chore: 용량 데이터 타입 변경

+++ src/main/java/com/tiki/server/document/entity/DeletedDocument.java
@@ -40,7 +40,7 @@ public class DeletedDocument extends BaseTime {
 	private long teamId;
 
 	@Column(nullable = false)
-	private double capacity;
+	private long capacity;
 
 	public static DeletedDocument of(final Document document) {
 		return DeletedDocument.builder()
