+++ src/main/java/com/tiki/server/timeblock/dto/response/TimeBlockDetailGetResponse.java
@@ -28,7 +28,7 @@ public record TimeBlockDetailGetResponse(
 		@NotNull long documentId,
 		@NotNull String fileName,
 		@NotNull String fileUrl,
-		@NotNull double capacity,
+		@NotNull long capacity,
 		@NotNull long tagId
 	) {
